+++ packages/gatsby/src/schema/__tests__/run-query.js
@@ -141,7 +141,7 @@ function makeGqlType(nodes) {
 
   const sc = createSchemaComposer()
   const typeName = `Test`
-  const tc = sc.createTC(typeName)
+  const tc = sc.createObjectTC(typeName)
   addInferredFields({
     schemaComposer: sc,
     typeComposer: tc,