+++ packages/gatsby/src/schema/infer/index.js
@@ -54,7 +54,7 @@ const addInferredTypes = ({
     let typeComposer
     let inferConfig
     if (schemaComposer.has(typeName)) {
-      typeComposer = schemaComposer.getTC(typeName)
+      typeComposer = schemaComposer.getOTC(typeName)
       inferConfig = getInferConfig(typeComposer)
       if (inferConfig.infer) {
         if (!typeComposer.hasInterface(`Node`)) {
@@ -62,22 +62,22 @@ const addInferredTypes = ({
         }
       }
     } else {
-      typeComposer = schemaComposer.createTC(typeName)
+      typeComposer = schemaComposer.createObjectTC(typeName)
       addNodeInterface({ schemaComposer, typeComposer })
     }
   })
 
   // XXX(freiksenet): We iterate twice to pre-create all types
-  const typeComposers = typeNames.map(typeName => {
+  const typeComposers = typeNames.map(typeName =>
     addInferredType({
       schemaComposer,
       nodeStore,
       typeConflictReporter,
-      typeComposer: schemaComposer.getTC(typeName),
+      typeComposer: schemaComposer.getOTC(typeName),
       typeMapping,
       parentSpan,
     })
-  })
+  )
 
   if (noNodeInterfaceTypes.length > 0) {
     noNodeInterfaceTypes.forEach(type => {